xtask: refactoring
