+++ xtask/src/build_man.rs
@@ -1,5 +1,13 @@
+use std::{
+    fs::{self, File},
+    iter,
+};
+
+use cargo_metadata::camino::{Utf8Path, Utf8PathBuf};
 use clap::{CommandFactory, Parser};
+use clap_mangen::Man;
 use color_eyre::eyre::Result;
+use time::OffsetDateTime;
 
 use crate::util;
 
@@ -18,11 +26,102 @@ impl Args {
         let cmd = souko::Args::command();
         let package_name = "souko";
         let section = "1";
-        for path in util::build_man_pages(&man_dir, cmd, package_name, section)? {
+        for path in build_man_pages(&man_dir, cmd, package_name, section)? {
             let path = path?;
             tracing::info!("  {}", util::to_relative(&path));
         }
 
         Ok(())
     }
+}
+
+fn build_man_pages<'a>(
+    man_dir: &Utf8Path,
+    mut cmd: clap::Command<'a>,
+    package_name: &str,
+    section: impl Into<String>,
+) -> Result<impl Iterator<Item = Result<Utf8PathBuf>> + 'a> {
+    cmd._build_all(); // https://github.com/clap-rs/clap/discussions/3603
+
+    let capitalized_name = {
+        let mut cs = package_name.chars();
+        match cs.next() {
+            Some(c) => c.to_uppercase().collect::<String>() + cs.as_str(),
+            None => String::new(),
+        }
+    };
+    let section = section.into();
+    let out_dir = man_dir.join(format!("man{section}"));
+    fs::create_dir_all(&out_dir)?;
+
+    let now = OffsetDateTime::now_utc();
+    let manual_name = format!("{capitalized_name} Command Manual");
+    let date = format!(
+        "{:04}-{:02}-{:02}",
+        now.year(),
+        u8::from(now.month()),
+        now.day()
+    );
+    let source = format!(
+        "{} {}",
+        cmd.get_name(),
+        cmd.get_version()
+            .or_else(|| cmd.get_long_version())
+            .unwrap_or_default()
+    );
+
+    let it = iterate_commands(cmd).map(move |cmd| {
+        let command_name = cmd.get_name().to_string();
+        let filename = format!("{command_name}.{}", section);
+        let path = out_dir.join(&filename);
+        let mut file = File::create(&path)?;
+        let man = Man::new(cmd.clone())
+            .title(command_name.to_uppercase())
+            .section(&section)
+            .date(&date)
+            .source(&source)
+            .manual(&manual_name);
+        man.render(&mut file)?;
+        Ok(path)
+    });
+
+    Ok(it)
+}
+
+fn iterate_commands<'a>(
+    cmd: clap::Command<'a>,
+) -> Box<dyn Iterator<Item = clap::Command<'a>> + 'a> {
+    #[allow(clippy::needless_collect)]
+    let subcommands = cmd.get_subcommands().cloned().collect::<Vec<_>>();
+    let command_name = cmd.get_name().to_string();
+    let command_version = cmd.get_version();
+    let command_long_version = cmd.get_long_version();
+    let command_author = cmd.get_author();
+
+    let it = iter::once(cmd.clone()).chain(
+        subcommands
+            .into_iter()
+            .map(move |mut subcommand| {
+                let name = format!("{command_name}-{}", subcommand.get_name());
+                subcommand = subcommand.name(name);
+                if subcommand.get_version().is_none() {
+                    if let Some(version) = command_version {
+                        subcommand = subcommand.version(version);
+                    }
+                }
+                if subcommand.get_long_version().is_none() {
+                    if let Some(long_version) = command_long_version {
+                        subcommand = subcommand.long_version(long_version);
+                    }
+                }
+                if subcommand.get_author().is_none() {
+                    if let Some(author) = command_author {
+                        subcommand = subcommand.author(author);
+                    }
+                }
+                subcommand
+            })
+            .flat_map(iterate_commands),
+    );
+    Box::new(it)
 }
